rebasing from master and fixing errors
